Fix magazine unlock at i-th event
During my playthough, I received the magazine supposed to be received in event #30 in event #28 because event #28 and #29 can be played out of order, and I did play #29 before #28. Since the magazine unlock code checks for the event before the magazine event (i.e. #29), the patch flagged the magazine unlock before the correct time